% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsconnect-package.R
\docType{package}
\name{dsconnect-package}
\alias{dsconnect}
\alias{dsconnect-package}
\title{dsconnect: Connects datasketch to R}
\description{
Connects datasketch to R
}
\author{
\strong{Maintainer}: Jp Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
