% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdb_create.R
\name{dsdb_create}
\alias{dsdb_create}
\title{Create a Database Object or Data Frames in a Database}
\usage{
dsdb_create(
  b,
  slug = NULL,
  name = NULL,
  description = NULL,
  org = NULL,
  authOrg = NULL,
  authToken = NULL
)
}
\arguments{
\item{b}{An object of type \code{hdbase} or a named list of data frames to create in the database.}

\item{slug}{A character string specifying the slug identifier for the database. Default is NULL.}

\item{name}{A character string specifying the name of the database. Default is NULL.}

\item{org}{A character string specifying the organization associated with the database. Default is NULL.}

\item{authOrg}{A character string specifying the organization to use for authentication. Default is NULL.}

\item{authToken}{A character string specifying the authentication token. Default is NULL.}
}
\value{
The newly created database object or the result of the creation of data frames in the database.
}
\description{
This function creates a new database object of type \code{hdbase} or, if 'b' is a named list of
data frames, it creates these data frames in the specified database. The operation requires
authentication details (organization and auth token).
}
\examples{
\dontrun{
# Assuming 'b' as a hdbase object or a named list of data frames,
# org name 'org_name', authentication org 'auth_org' and authentication token 'auth_token'
result <- dsdb_create(b = b, org = 'org_name', authOrg = 'auth_org', authToken = 'auth_token')
}
}
