% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsdb_read.R
\name{dsdb_read}
\alias{dsdb_read}
\title{Read Data Frames from a Database}
\usage{
dsdb_read(db, org = NULL, authOrg = NULL, authToken = NULL)
}
\arguments{
\item{db}{A database slug (character string) specifying the database from which to read data frames.}

\item{org}{A character string specifying the organization associated with the database. Default is NULL.}

\item{authOrg}{A character string specifying the organization to use for authentication. Default is NULL.}

\item{authToken}{A character string specifying the authentication token. Default is NULL.}
}
\value{
A data frame containing the requested data from the specified database.
}
\description{
This function retrieves data frames from a specified database in datasketch.
It requires authentication details (organization and auth token).
}
\examples{
\dontrun{
# Assuming a valid database connection 'db_conn', org name 'org_name',
# authentication org 'auth_org' and authentication token 'auth_token'
data <- dsdb_read(db = 'db_slug', org = 'org_name', authOrg = 'auth_org', authToken = 'auth_token')
}
}
