% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsviz_list.R
\name{dsviz_list}
\alias{dsviz_list}
\title{List Visualizations}
\usage{
dsviz_list(org = NULL, authOrg = NULL, authToken = NULL)
}
\arguments{
\item{org}{A character string specifying the organization associated with the visualizations. Default is NULL.}

\item{authOrg}{A character string specifying the organization to use for authentication. Default is NULL.}

\item{authToken}{A character string specifying the authentication token. Default is NULL.}
}
\value{
A list of visualizations associated with the specified organization.
}
\description{
This function retrieves a list of visualizations available to the specified organization
in the Datasketch Cloud Hosting environment. The operation requires authentication details
(organization and auth token).
}
\examples{
\dontrun{
# Assuming org name 'org_name', authentication org 'auth_org' and authentication token 'auth_token'
viz_list <- dsviz_list(org = 'org_name', authOrg = 'auth_org', authToken = 'auth_token')
}
}
